+++ lib/dynamic.dart
@@ -1,0 +1,6 @@
+library liquid.dynamic;
+
+import 'dart:mirrors';
+import 'package:liquid/liquid.dart';
+
+part 'package:liquid/src/dynamic/component_factory.dart';